Move file iterator to separate file
